Restructured into a python package, added basic setup.py
