+++ pykinect2/PyKinectRuntime.py
@@ -1,5 +1,5 @@
-import PyKinectV2
-from PyKinectV2 import *
+from pykinect2 import PyKinectV2
+from pykinect2.PyKinectV2 import *
 
 import ctypes
 import _ctypes
