+++ pykinect2/examples/PyKinectBodyGame.py
@@ -1,6 +1,6 @@
-import PyKinectV2
-from PyKinectV2 import *
-import PyKinectRuntime
+from pykinect2 import PyKinectV2
+from pykinect2.PyKinectV2 import *
+from pykinect2 import PyKinectRuntime
 
 import ctypes
 import _ctypes
